For Whom The Bell Tolls
Level 1 Puzzle Design complete
Functionality - just about

Player collides with triggerbox which then allows them to switch to "Puzzle mode" disabling player movement and switching the camera for a better view.

On screen current controls for completing the puzzle - not final

Objective
Get a ball down each pipe once to activate the battery and unlock the door entrance to go back to the hub to complete the level. SICK. CONTENT.

Waypoints set up to Guide the player

Doors work. ( apart from the ones that dont)

More to come...