Build Before Milestone 7
This build was focused on cleaning up the game for submission.
The Tutorial stage features no wall jumping bugs, as well as ambient audio for the Station, as well as sound effects.
We need to discuss some major build flaws, i'll try to have a session with Vik/David before Class